style: clean code
