+++ src/main/java/org/opensourceway/sbom/manager/batch/reader/license/ExtractLicenseReader.java
@@ -5,12 +5,12 @@ import org.apache.commons.collections4.ListUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.opensourceway.sbom.constants.BatchContextConstants;
-import org.opensourceway.sbom.manager.batch.processor.license.ExtractLicensesProcessor;
 import org.opensourceway.sbom.manager.dao.SbomRepository;
 import org.opensourceway.sbom.manager.model.ExternalPurlRef;
 import org.opensourceway.sbom.manager.model.License;
 import org.opensourceway.sbom.manager.model.Package;
 import org.opensourceway.sbom.manager.model.Sbom;
+import org.opensourceway.sbom.manager.service.license.LicenseService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.ChunkListener;
@@ -39,8 +39,7 @@ public class ExtractLicenseReader implements ItemReader<List<ExternalPurlRef>>, 
     private static final Logger logger = LoggerFactory.getLogger(ExtractLicenseReader.class);
 
     @Autowired
-//    @Qualifier("licenseServiceImpl")
-    private ExtractLicensesProcessor extractLicensesProcessor;
+    private LicenseService licenseService;
 
     @Autowired
     SbomRepository sbomRepository;
@@ -54,7 +53,7 @@ public class ExtractLicenseReader implements ItemReader<List<ExternalPurlRef>>, 
     private ChunkContext chunkContext;
 
     private void initMapper(UUID sbomId) {
-        if (!extractLicensesProcessor.needRequest()) {
+        if (!licenseService.needRequest()) {
             logger.warn("license client does not request");
             return;
         }
@@ -75,7 +74,7 @@ public class ExtractLicenseReader implements ItemReader<List<ExternalPurlRef>>, 
                 .filter(externalPurlRef -> externalPurlRef.getCategory().equals("PACKAGE_MANAGER"))
                 .toList();
 
-        this.chunks = ListUtils.partition(externalPurlRefs, extractLicensesProcessor.getBulkRequestSize())
+        this.chunks = ListUtils.partition(externalPurlRefs, licenseService.getBulkRequestSize())
                 .stream()
                 .map(ArrayList::new)// can't use ArrayList.subList(can't restart)
                 .collect(Collectors.toCollection(CopyOnWriteArrayList::new));// can't use unmodifiableList(can't remove)
